feat: PV再構築機能を汎用化し、sharded_ttとttモジュールにおける再構築処理を統一。TTProbeトレイトを実装
